requirements.txt
asyncpg==0.27.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.8.22
GeoAlchemy2==0.12.5
geopandas==0.12.1
greenlet==2.0.1
idna==3.4
loguru==0.6.0
lxml==4.9.2
munch==2.5.0
numpy==1.23.5
packaging==21.3
pandas==1.5.2
psycopg2-binary==2.9.5
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
Rtree==1.0.1
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
asyncpg==0.27.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.8.22
Flask==2.2.2
GeoAlchemy2==0.12.5
geopandas==0.12.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.1
munch==2.5.0
numpy==1.23.5
packaging==21.3
pandas==1.5.2
psycopg2-binary==2.9.5
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
Rtree==1.0.1
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
